submissions/task-4.3/submission-4.3/agent/requirements.txt
openai>=1.40.0
PyYAML>=6.0
cocotb>=1.8
